Remove .eslintrc.yml from workflows folder (#62)
The .eslintrc.yml file is unnecessary since it was moved to the root directory of the repo. The file also confuses GitHub and makes it think that .eslintrc.yml is a workflow. This causes GitHub to run the .yml file; however, since it is not a workflow, but rather a config file for the eslint linter, it fails as a test.